ugly, yet working cpp

+++ kernel/inc/types.h
@@ -2,6 +2,7 @@
 // Copyright (c) 2025 0xKarinyash
 
 #pragma once
+#include <stddef.h>
 
 typedef unsigned char      u8;
 typedef unsigned short     u16;
@@ -13,7 +14,7 @@ typedef signed short       i16;
 typedef signed int         i32;
 typedef signed long long   i64;
 
-typedef u64 usize;
+typedef size_t usize;
 typedef u64 uintptr_t;
 
 typedef struct {
